Fixing ExchangeRateHistory.java method
