+++ src/main/java/pl/isa/javaers/DailyAlertChecker.java
@@ -10,9 +10,9 @@ import java.util.Scanner;
 
 public class DailyAlertChecker {
     private String kursNBP = new Rates().getTable();                                                           // dane z pliku json pobrane do porównania
-    private String rateEffectiveDate = String.valueOf(new Rate().getEffectiveDate());
+    private String rateEffectiveDate = String.valueOf(new CurrRate().getEffectiveDate());
     private String rateCurrency = new Rates().getCode();
-    private float rateCourse = new Rate().getMid();//String.valueOf(new Rate().getMid());
+    private float rateCourse = new CurrRate().getMid();//String.valueOf(new Rate().getMid());
     private String alertID = String.valueOf(new Alert().getAlertID());                                          // dane z alertu ustawionego przez uzytkownika pobrane do porownania
     private float userCourse = new Alert().getCourse();  //String.valueOf(new Alert().getCourse());
     private String userCurrency = new Alert().getCurrCode();
