+++ src/main/java/pl/isa/javaers/RatesParser.java
@@ -8,28 +8,24 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.*;
-import java.util.logging.Handler;
 
 public class RatesParser {
 
     @JsonIgnoreProperties(ignoreUnknown = true)
-    public static HashMap<Rates, Rate> getListOfRatesFromJSON(String currCode) {
+    public static List<CurrRate> listFromJSON(String currCode) {
 
         try {
             byte[] data = Files.readAllBytes(Paths.get("src/main/resources/KursyNBP/" + currCode + ".json"));
             ObjectMapper objectMapper = new ObjectMapper();
             objectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-//            return Arrays.asList(objectMapper.readValue(data, Rates[].class));
-            HashMap<Rates, Rate> rateMap = new HashMap<>();
-rateMap.putAll((Map<? extends Rates, ? extends Rate>) objectMapper.readValue(data, Rates.class));
-            return rateMap;
+//            List<Rates> listFromJSON = Arrays.asList(objectMapper.readValue(data, Rates[].class));
+//            return listFromJSON;
+
+//            return Collections.singletonList(objectMapper.readValue(data, CurrRate.class));
+            return Arrays.asList(objectMapper.readValue(data, CurrRate.class));
+//            HashMap<Rates, Rate> rateMap = new HashMap<>();
         } catch (IOException e) {
             throw new RuntimeException(e);
-//            return new ArrayList<>();
         }
     }
-//    public static List<Rate> loadRate(String currCode){
-//        List<Rate> rates = getListOfRatesFromJSON(currCode);
-//        return rates;
-//    }
 }
